requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
openai>=1.57.0
anthropic>=0.39.0
google-generativeai>=0.8.0
langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0,<1.0.0
langgraph>=0.2.0,<1.0.0
langsmith>=0.1.0
dspy>=2.5.0
sentence-transformers>=3.3.0
huggingface-hub>=0.26.0
chromadb>=0.5.0
qdrant-client>=1.12.0
pypdf>=5.1.0
pymupdf>=1.25.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
httpx>=0.28.0
aiohttp>=3.11.0
feedparser>=6.0.0
trafilatura>=2.0.0
pandas>=2.2.0
numpy>=2.0.0
tenacity>=9.0.0
tiktoken>=0.8.0
pyyaml>=6.0.2
structlog>=24.4.0
rich>=13.9.0
asyncio-throttle>=1.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
